refactor: do not use `Item` for equipments
Using `Item` for the equipment allows actors to equip non-equipable
items, like herbs and books.  Firstly I tried to replace `Effect` with a
type variable `a`, and define the item type as `Item a` and inventory as
`forall a. Effect a => [a]`. However, using `forall` is difficult to
handle. Also, we know `a` can be `Herb`, `Book`, `Weapon`, or `Armor`.
And I found `open-union`. It was what I wanted.