Use only the filename as the map key
The difference in path styles between Windows and *nix caused some tests
to fail. However, there was no library to absorb the differences, so I
changed the code to use only the filename as the map key. It is not
intuitive, so later, I may revert and write some codes to convert
Windows-style paths to *nix-style ones.