Specify paths that actually exist
The previous caching method did not work on Windows because the paths
were wrong. This commit specifies the `.cabal/store` directory path
based on the output of `haskell/actions/setup`. We no longer save the
contents of `.cabal/packages`, but I think it is fine because the
`packages` directory only contains packages' tarballs.