Expose `addTileFile`
`readTileFileRecursive` is difficult to test. We must create all the expected tile structures even if we do not test some of them. Especially, it is difficult to create a test case for generation tiles because we will change their structure, and generation tiles will have 47+ tiles.
That is why we expose `addTileFile`.